requirements_cloud.txt
streamlit>=1.28.0
pandas>=1.5.0
Pillow>=9.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
